requirements.txt
openai>=1.0.0
httpx>=0.27.0
flake8>=7.0.0
pyflakes>=3.0.0
fawltydeps>=0.15.0
mypy>=1.9.0
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-dotenv
python-multipart
websockets
